wip - do not merge
This commit was to experiment with the name of the file that we serve/and build
and whether we could encode the source file directly the URL ('/App.svelte.js').

This worked well in a lot of ways, but ran up against a big issue with users
partial-importing files. Right now, we leverage the fact that a user writes
"import './App'" in their source code, and it almost always means a file that builds
to "js", so we can naively add the missing ".js" file extension and call it a day.

In this commit, we would need to look up multiple different files on disk to
resolve to the right one.