Don't prefer built-in Node.js modules when running `esinstall`
If a module with the same name as a built-in Node.js module exists in
`node_modules`, it should take precedence over the built-in module. For
example, if the `buffer` module from npm is used by a package,
`esinstall` should use this module instead of failing with an error
message saying that Node.js polyfills are needed.