Note to users that HMR is not fully supported yet
Just a small note that HMR is not fully supported yet for `snowpack build` - I was really excited reading that it was and spent quite some time trying to get it working before finding the github thread. Hopefully this will inform other users who are excited for this HMR approach to soon be working!