Pin rollup version to fix tree shaking bugs (#2572)
Recent versions of rollup have started tree shaking by removing bindings while leaving the assignment expressions in the code. This breaks `lottie-web` or any package which binds to a function like this:

```js
var foo = function bar() {

}.bind(this);
```

This is being compiled to:

```js
function bar() {

}.bind(this);
```

There might be other cases where removing the binding is also not safe.

Solution is to pin to the previous working version of rollup.