Add resolve package to snowpack direct dependencies
Uses of the `resolve` package are sometimes getting bundled into snowpack, but this isn't listed as a direct dependency.

This is a common package, so I suspect for most users it is resolvable. However, for new projects (e.g. the minimal create-snowpack-app using snowpack v3.0.12) this is isn't getting installed, causing a module not found error when running snowpack.