Remove enumerability of converted ESM in SSR (#2920)
This fixes #2918

The behavior of Rollup's CommonJS plugin to allow CommonJS modules to treat the default export as a function relied on named exports being non-enumerable. This was regressed in 689f58e to fix another bug related to re-exporting everything from another module.

That code was using enumerability to re-export. We can fix both cases by not using enumerability in the export star code, but rather grabbing the property descriptors.