Update jest.md
Unless I am missing something, I struggled mightily to get Jest to work with TS and JSX until I came back to this page for the 10th-or-so time and thought maybe there was a typo in the name of the necessary root-level config file. This seems to have fixed all of my issues.