Use a separate PackageSource for each config
In 3.3.0 we changed how Snowpack config is passed around through the PackageSource. Now that gets passed in as part of the constructor and not through the instance methods. The problem is that these PackageSources are cached globally. So in an environment with multiple snowpack instances it would not be able to resolve packages due to a global single PackageSource.

This fixes that, making PackageSources be distinct by snowpack config.