Expand devOptions.open documentation
## Changes
<!-- What does this change, in plain language? -->
<!-- Before/after screenshots may be helpful.  -->
Adds a snippet to the description of the devOptions.open config that explains you can also use a path to specify which browser to open

## Testing

<!-- How was this change tested? -->
<!-- DON'T DELETE THIS SECTION! If no tests added, explain why -->
No, this is a change to documentation

## Docs
Add note that path is a valid input to devOptions.open
<!-- Was public documentation updated? -->
<!-- DON'T DELETE THIS SECTION! If no docs added, explain why (e.g. "bug fix only") -->
Yes, noted in change. Better describes the allowed input to the config option.