Ensure remote package fetches always use the configured origin
If a specific `packageOptions.origin` value is provided when using remote streaming, the origin was _**not**_ always used when fetching resources, because `snowpack/src/util/ts` had a hard-coded refernce to `https://pkg.snowpack.dev` rather than using the config value.

This PR ensures that the `SkypackSDK` instances are _always_ created with the project's snowpack configuration.  It also removes some duplication of the strings that define the two remote streaming origins, just for code cleanliness.  Also, tests (such as they are) have been added to verify that remote streaming is done from `pkg.snowpack.dev` by default, or by the configured origin.

Provides a workaround for #3393 and skypackjs/skypack-cdn#179.