Fix race condition reading existing import-map.json files (#3453)
* Fix race condition reading existing import-map.json files

When trying to read the import-map.json, if that file is being concurrently using `fs.stat` won't tell you that. Using `fs.open` gives you a handle with `r+` (read and write) permission, which throws if it's open somewhere else for writing. If not, you get a handle and can read the data.

Fixes https://github.com/snowpackjs/astro/issues/468

* Tighten the logic even more