Fix doctype code block in documentation (#3506)
The inline `<!doctype html>` mention was being interpreted as literal HTML which was then (presumably) being sanitised by the tool building. This issue isn't apparent when viewing the markdown in Github because github just escapes and preserves the HTML.

I wasn't able to build the documentation locally to verify this fixed it, because I don't have astro (??) installed locally (??), but I think this should do the trick.