Make App Scripts React compatible with Jest 27 (#3750)
* Upgrade babel-jest version

This is to match with the current version of Jest. Also Jest is added
as a peer dependency to reflect this.

* Change the babel jest transform script

This is done to reflect the new API changes in babel-jest.
The second line is done so that it can stay backwards compatible with
Jest 26 if needed.

References:
https://github.com/snowpackjs/snowpack/issues/3398#issuecomment-887499318
https://github.com/facebook/jest/issues/11444#issuecomment-855989054

* Use jsdom as a test environment to run jest

This is done so the react testing-library can render the DOM elements
in testing.