Fix 404s in dev and "Module not found" in production for TS files containing only types
- Fixes https://github.com/withastro/snowpack/issues/3285
- Fixes https://github.com/withastro/snowpack/issues/3585

When building a TS file with only types, esbuild strips out all of the type info and returns an *empty string* as the result of the compilation (`''`).
`loadUrl()` just does a falsy check on the return value from `fileBuilder.getResult`, which can be any of `string | Buffer | undefined`.
When the empty string is returned, the falsy check doesn't pass, so the load appears to be unsuccessful even though the empty string indicates successful output.
Changing the falsy check to an explicit check for `undefined` allows the empty string to be treated as a successful response.

An alternative to this approach might be to just ignore the import entirely (somehow) if the response is empty. That would potentially save a bunch of useless requests from being sent. But this is at least a step in the right direction that helps code like this to be runnable at least.